Set up main framework
Set up the basis for the project with a semi-functioning
calendar/display function. Added a base room information text file for
reading in. Added a format guide for text files in order to describe the
format of other text files (i.e. what each piece of information
represents) for easier understanding.